requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
crawl4ai>=0.4.0
streamlit>=1.37.0
requests>=2.32.0
strands-agents>=1.0.0
strands-agents-tools>=0.1.0
bedrock-agentcore
bedrock-agentcore-starter-toolkit
openai>=1.40.0
boto3>=1.34.0
botocore>=1.34.0
pydantic>=2.8.0
pydantic-settings>=2.4.0
structlog>=24.4.0
python-dotenv>=1.0.0
beautifulsoup4>=4.12.0
python-slugify>=8.0.0
lxml>=5.0.0
